Users/ADMIN/Documents/NetBeansProjects/ExamenPIII/src/da/Jugador.java
package da;

/**
 *
 * @author ADMIN
 */
public class Jugador {

    private String nombre;
    private String apellido;
    private int edad;
    private String posicion;
    private boolean titular;

    public Jugador(String nombre, String apellido, int edad, String posicion, boolean titular) {

        this.nombre = nombre;
        this.apellido = apellido;
        this.edad = edad;
        this.posicion = posicion;
        this.titular = titular;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getApellido() {
        return apellido;
    }

    public void setApellido(String apellido) {
        this.apellido = apellido;
    }

    public int getEdad() {
        return edad;
    }

    public void setEdad(int edad) {
        this.edad = edad;
    }

    public String getPosicion() {
        return posicion;
    }

    public void setPosicion(String posicion) {
        this.posicion = posicion;
    }

    public boolean isTitular() {
        return titular;
    }

    public void setTitular(boolean titular) {
        this.titular = titular;
    }

}
